Refactored ProcessMeasurement method and fixed prediction
The time change between measurements was being incorrectly calculated.
Previously, the calculation was assuming that the time stamps were in
nanoseconds. In fact, they are in microseconds.

The ProcessMeasurement method was too long, so it was split up to
improve readability.

Disabled addition of the randomness that is supposed to model
acceleration. With it disabled, the inability of the model to
handle the figure 8 pattern is demonstrated. It seems that the
randomness is helpful, but its magnitude might be too large.