Test higher radar range rate variance
Increasing the uncertainty associated with the radar range rate by
increasing the variance entry in the radar measurement covariance
matrix results in a lower RMSE for vy and a higher RMSE for vx. It
seems that taking more information from the radar velocity measurements
results in an increase in error of vx estimation, and a decrease in
error of vy estimation. Why? Perhaps it is related to the shape of
the path. To understand why this is the case, it is necessary to
understand how the radar range rate information is incorporated into
the state estimation.

(0.0983, 0.0963, 0.5741, 0.4982)